MLK-21052-12 soc: imx: Redefine sc_rsrc_t to uint32_t from uint16_t
This was changed at some point by SC firmware but imx linux releases
didn't pick up the change and some drivers don't even compile with a
16-bit sc_rsrc_t.

Fix by overwriting SC API with imx_4.14 definition.

+++ include/soc/imx8/sc/types.h
@@ -824,7 +824,7 @@ typedef uint8_t sc_err_t;
  * and bus masters (but not memory regions). Note items from list should
  * never be changed or removed (only added to at the end of the list).
  */
-typedef uint16_t sc_rsrc_t;
+typedef uint32_t sc_rsrc_t;
 
 /*!
  * This type is used to indicate a control.
